Move tests out of the library crate, into a single test crate
