+++ crates/apollo-compiler/Cargo.toml
@@ -15,6 +15,7 @@ categories = [
   "web-programming",
 ]
 edition = "2021"
+autotests = false # Single executable for all tests
 
 [dependencies]
 apollo-parser = { path = "../apollo-parser", version = "0.6.0" }
@@ -41,3 +42,7 @@ pretty_assertions = "1.3.0"
 name = "multi-source"
 path = "benches/multi_source.rs"
 harness = false
+
+# Single executable for all tests
+[[test]]
+name = "tests"
