+++ crates/apollo-compiler/src/database/document.rs
@@ -370,183 +370,3 @@ pub(crate) fn is_subtype(
         .get(&abstract_type)
         .map_or(false, |set| set.contains(&maybe_subtype))
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::ApolloCompiler;
-    use crate::HirDatabase;
-
-    #[test]
-    fn find_operations() {
-        let type_system = r#"
-type Query {
-  name: String
-}
-        "#;
-        let op = r#"{ name }"#;
-        let named_op = r#"query getName { name } "#;
-        let several_named_op = r#"query getName { name } query getAnotherName { name }"#;
-        let noop = r#""#;
-
-        let mut compiler = ApolloCompiler::new();
-        compiler.add_type_system(type_system, "ts.graphql");
-        let op_id = compiler.add_executable(op, "op.graphql");
-        let op = compiler.db.find_operation(op_id, None);
-        assert!(op.is_some());
-
-        compiler.update_executable(op_id, named_op);
-        let op = compiler.db.find_operation(op_id, Some("getName".into()));
-        assert!(op.is_some());
-        let op = compiler.db.find_operation(op_id, None);
-        assert!(op.is_some());
-
-        compiler.update_executable(op_id, several_named_op);
-        let op = compiler.db.find_operation(op_id, Some("getName".into()));
-        assert!(op.is_some());
-        let op = compiler.db.find_operation(op_id, None);
-        assert!(op.is_none());
-
-        compiler.update_executable(op_id, noop);
-        let op = compiler.db.find_operation(op_id, Some("getName".into()));
-        assert!(op.is_none());
-
-        let op = compiler.db.find_operation(op_id, None);
-        assert!(op.is_none());
-    }
-
-    #[test]
-    fn find_definitions_with_directive() {
-        let schema = r#"
-            type ObjectOne @key(field: "id") {
-              id: ID!
-              inStock: Boolean!
-            }
-
-            type ObjectTwo @key(field: "name") {
-              name: String!
-              address: String!
-            }
-
-            type ObjectThree {
-                price: Int
-            }
-        "#;
-
-        let mut compiler = ApolloCompiler::new();
-        compiler.add_document(schema, "schema.graphql");
-
-        let key_definitions = compiler.db.find_types_with_directive(String::from("key"));
-        let mut key_definition_names: Vec<&str> =
-            key_definitions.iter().map(|def| def.name()).collect();
-        key_definition_names.sort();
-        assert_eq!(key_definition_names, ["ObjectOne", "ObjectTwo"])
-    }
-
-    #[test]
-    fn is_subtype() {
-        fn gen_schema_types(schema: &str) -> ApolloCompiler {
-            let base_schema = with_supergraph_boilerplate(
-                r#"
-                type Query {
-                me: String
-                }
-                type Foo {
-                me: String
-                }
-                type Bar {
-                me: String
-                }
-                type Baz {
-                me: String
-                }
-
-                union UnionType2 = Foo | Bar
-                "#,
-            );
-            let schema = format!("{base_schema}\n{schema}");
-            let mut compiler = ApolloCompiler::new();
-            compiler.add_document(&schema, "schema.graphql");
-            compiler
-        }
-
-        fn gen_schema_interfaces(schema: &str) -> ApolloCompiler {
-            let base_schema = with_supergraph_boilerplate(
-                r#"
-                type Query {
-                me: String
-                }
-                interface Foo {
-                me: String
-                }
-                interface Bar {
-                me: String
-                }
-                interface Baz {
-                me: String,
-                }
-
-                type ObjectType2 implements Foo & Bar { me: String }
-                interface InterfaceType2 implements Foo & Bar { me: String }
-                "#,
-            );
-            let schema = format!("{base_schema}\n{schema}");
-            let mut compiler = ApolloCompiler::new();
-            compiler.add_document(&schema, "schema.graphql");
-            compiler
-        }
-
-        let ctx = gen_schema_types("union UnionType = Foo | Bar | Baz");
-        assert!(ctx.db.is_subtype("UnionType".into(), "Foo".into()));
-        assert!(ctx.db.is_subtype("UnionType".into(), "Bar".into()));
-        assert!(ctx.db.is_subtype("UnionType".into(), "Baz".into()));
-
-        let ctx =
-            gen_schema_interfaces("type ObjectType implements Foo & Bar & Baz { me: String }");
-        assert!(ctx.db.is_subtype("Foo".into(), "ObjectType".into()));
-        assert!(ctx.db.is_subtype("Bar".into(), "ObjectType".into()));
-        assert!(ctx.db.is_subtype("Baz".into(), "ObjectType".into()));
-
-        let ctx = gen_schema_interfaces(
-            "interface InterfaceType implements Foo & Bar & Baz { me: String }",
-        );
-        assert!(ctx.db.is_subtype("Foo".into(), "InterfaceType".into()));
-        assert!(ctx.db.is_subtype("Bar".into(), "InterfaceType".into()));
-        assert!(ctx.db.is_subtype("Baz".into(), "InterfaceType".into()));
-
-        let ctx = gen_schema_types("extend union UnionType2 = Baz");
-        assert!(ctx.db.is_subtype("UnionType2".into(), "Foo".into()));
-        assert!(ctx.db.is_subtype("UnionType2".into(), "Bar".into()));
-        assert!(ctx.db.is_subtype("UnionType2".into(), "Baz".into()));
-
-        let ctx = gen_schema_interfaces("extend type ObjectType2 implements Baz { me2: String }");
-        assert!(ctx.db.is_subtype("Foo".into(), "ObjectType2".into()));
-        assert!(ctx.db.is_subtype("Bar".into(), "ObjectType2".into()));
-        assert!(ctx.db.is_subtype("Baz".into(), "ObjectType2".into()));
-
-        let ctx =
-            gen_schema_interfaces("extend interface InterfaceType2 implements Baz { me2: String }");
-        assert!(ctx.db.is_subtype("Foo".into(), "InterfaceType2".into()));
-        assert!(ctx.db.is_subtype("Bar".into(), "InterfaceType2".into()));
-        assert!(ctx.db.is_subtype("Baz".into(), "InterfaceType2".into()));
-    }
-
-    fn with_supergraph_boilerplate(content: &str) -> String {
-        format!(
-            "{}\n{}",
-            r#"
-            schema
-                @core(feature: "https://specs.apollo.dev/core/v0.1")
-                @core(feature: "https://specs.apollo.dev/join/v0.1") {
-                query: Query
-            }
-            directive @core(feature: String!) repeatable on SCHEMA
-            directive @join__graph(name: String!, url: String!) on ENUM_VALUE
-            enum join__Graph {
-                TEST @join__graph(name: "test", url: "http://localhost:4001/graphql")
-            }
-
-            "#,
-            content
-        )
-    }
-}
