+++ crates/apollo-compiler/tests/executable.rs
@@ -1,5 +1,44 @@
-use crate::ApolloCompiler;
-use crate::ReprDatabase;
+use apollo_compiler::ApolloCompiler;
+use apollo_compiler::HirDatabase;
+use apollo_compiler::ReprDatabase;
+
+#[test]
+fn find_operations() {
+    let type_system = r#"
+type Query {
+name: String
+}
+    "#;
+    let op = r#"{ name }"#;
+    let named_op = r#"query getName { name } "#;
+    let several_named_op = r#"query getName { name } query getAnotherName { name }"#;
+    let noop = r#""#;
+
+    let mut compiler = ApolloCompiler::new();
+    compiler.add_type_system(type_system, "ts.graphql");
+    let op_id = compiler.add_executable(op, "op.graphql");
+    let op = compiler.db.find_operation(op_id, None);
+    assert!(op.is_some());
+
+    compiler.update_executable(op_id, named_op);
+    let op = compiler.db.find_operation(op_id, Some("getName".into()));
+    assert!(op.is_some());
+    let op = compiler.db.find_operation(op_id, None);
+    assert!(op.is_some());
+
+    compiler.update_executable(op_id, several_named_op);
+    let op = compiler.db.find_operation(op_id, Some("getName".into()));
+    assert!(op.is_some());
+    let op = compiler.db.find_operation(op_id, None);
+    assert!(op.is_none());
+
+    compiler.update_executable(op_id, noop);
+    let op = compiler.db.find_operation(op_id, Some("getName".into()));
+    assert!(op.is_none());
+
+    let op = compiler.db.find_operation(op_id, None);
+    assert!(op.is_none());
+}
 
 #[test]
 fn is_introspection_operation() {
