+++ crates/apollo-compiler/tests/tests.rs
@@ -1,0 +1,8 @@
+mod executable;
+mod merge_schemas;
+/// Formerly in src/lib.rs
+mod misc;
+mod parser;
+mod schema;
+mod snapshot_tests;
+mod validation;
